Css working. Needs to be organized
