--- portfolio/src/App.css
@@ -1,106 +1,0 @@
-body {
-  background-color: black;
-  padding: 1%;
-  min-height: 100vh;
-}
-
-.App {
-  text-align: center;
-  background-color: lightgreen;
-  padding: 2%;
-}
-
-nav {
-  padding: 50px;
-}
-
-a {
-  text-decoration: none;
-  color: black;
-}
-
-nav a {
-  padding: 25px 45px;
-  text-decoration: none;
-  font-size: 2.5em;
-}
-
-input {
-  display: flex;
-  margin: auto;
-  margin-bottom: 10px;
-  width: 200px;
-  padding: 5px 10px;
-  border-radius: 5px;
-  /*box-shadow: 2px 2px 2px grey;*/
-  border: 2px double black;
-}
-
-.trip {
-  /*display: flex;*/
-  /*flex-direction: column;*/
-  /*justify-content: left;*/
-  text-align: left;
-  width: 50%;
-  border: 2px solid black;
-  margin: 2% auto;
-  padding: 2%;
-  background-color: lightblue;
-  border-radius: 10px;
-  color: black;
-}
-
-.trip h1 {
-  text-align: center;
-}
-
-.CreateTrip {
-  width: 50%;
-  margin: auto;
-
-}
-
-form {
-  margin: auto;
-  width: 60%;
-  padding: 3%;
-  border-radius: 20px;
-  border: 2px solid black;
-  background-color: lightgrey;
-}
-
-input .checkbox {
-  /*width: 60px;*/
-}
-
-.CreateTrip h1 {
-  width: 50%;
-  /*margin-bottom: 2%;*/
-}
-
-button {
-  margin-left: 75%;
-  padding: 1% 3%;
-  border-radius: 5px;
-  background-color: silver;
-  margin-bottom: 3%;
-}
-
-form button {
-  margin-left: 40%;
-  padding: 2% 8%;
-  border-radius: 5px;
-  background-color: silver;
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
