+++ portfolio/src/css/App.css
@@ -97,6 +97,12 @@ input {
 
 .trip h1 {
   text-align: center;
+  padding: 20px 10px 20px 10px;
+  background-color: #e7e7e7;
+  margin: 0 auto;
+  font-size: 1rem;
+  border-top-left-radius: 10px;
+  border-top-right-radius: 10px;
 }
 
 .CreateTrip {
@@ -120,6 +126,12 @@ input .checkbox {
 
 .CreateTrip h1 {
   width: 50%;
+  padding: 20px 10px 20px 10px;
+  background-color: #e7e7e7;
+  margin: 0 auto;
+  font-size: 1rem;
+  border-top-left-radius: 10px;
+  border-top-right-radius: 10px;
   /*margin-bottom: 2%;*/
 }
